+++ stroom-search/stroom-search-impl/src/main/java/stroom/search/impl/AsyncSearchTaskHandler.java
@@ -19,7 +19,11 @@ package stroom.search.impl;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import stroom.cluster.task.api.*;
+import stroom.cluster.task.api.ClusterDispatchAsync;
+import stroom.cluster.task.api.ClusterTaskTerminator;
+import stroom.cluster.task.api.NodeNotFoundException;
+import stroom.cluster.task.api.NullClusterStateException;
+import stroom.cluster.task.api.TargetNodeSetFactory;
 import stroom.index.impl.IndexShardService;
 import stroom.index.impl.IndexStore;
 import stroom.index.shared.FindIndexShardCriteria;
@@ -37,8 +41,13 @@ import stroom.util.shared.Sort.Direction;
 
 import javax.inject.Inject;
 import javax.inject.Provider;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.Map.Entry;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 class AsyncSearchTaskHandler {